feat(protocol): update protocol from 14.12 to 15.11
- Update client version to 15.11 (1511)
- Rewrite onConnect challenge packet format (U8 length, 0x71 extra byte)
- Rewrite onRecvFirstMessage for 15.11 (asset hash, OTCv8 detection)
- Complete AddPlayerSkills rewrite (percent*100, combat stats as doubles)
- Reorder login packet sequence to match 15.11 protocol
- Add new stub packets (bosstiary, prey, forge, proficiency, VIP groups, etc.)
- Update MagicEffectClasses/ShootType_t from uint8_t to uint16_t throughout
- Update wire format to use uint16_t for magic/distance effects in loop system
- Add new enums: InspectObjectType, Imbuement_Window_t, WeaponProficiency_t
- Expand NETWORKMESSAGE_MAXSIZE to 65500, Icons_t, ResourceTypes_t

+++ src/combat.cpp
@@ -327,12 +327,12 @@ bool Combat::setParam(CombatParam_t param, uint32_t value)
 		}
 
 		case COMBAT_PARAM_EFFECT: {
-			params.impactEffect = static_cast<uint8_t>(value);
+			params.impactEffect = static_cast<uint16_t>(value);
 			return true;
 		}
 
 		case COMBAT_PARAM_DISTANCEEFFECT: {
-			params.distanceEffect = static_cast<uint8_t>(value);
+			params.distanceEffect = static_cast<uint16_t>(value);
 			return true;
 		}
